requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==22.0.0
pandas==2.2.2
openpyxl==3.1.5
rapidfuzz==3.9.6
python-dateutil==2.9.0.post0
plotly==5.24.1
xlrd==1.2.0
